+++ src/components/CenterInfo/CenterInfoType.ts
@@ -22,3 +22,35 @@ export interface IWayToComeResponseType {
   parkinginfoName: string;
   parkinginfoContent: string;
 }
+
+export interface ISodaInfoType {
+  title: string;
+  headText: string;
+  img: string;
+  alt: string;
+  content: Array<string>;
+  detailImg: Array<string>;
+}
+
+export interface IReservationInfoType {
+  title: string;
+  reservation: Array<IReservation>;
+  counselingTime: Array<ICounselingTimeType>;
+}
+export interface ICounselingTimeType {
+  name: string;
+  time: string;
+}
+
+export interface IReservation {
+  img: string;
+  step: number;
+  stepName: string;
+  stepInfo: string;
+}
+
+export interface INavListItemMenuType {
+  url: string;
+  listName: string;
+  listItem: Array<string>;
+}
